fix(sound): commands are ignored if channel muted
Bug:
Any commands of a channel are skipped and ignored when that channel is
muted, even if a command was relevant to `volume` or `envelope`. This
causes incorrect volume setting of background music if it was muted
explicitly (by user) or implicitly (by sound effects). Typically, if
background music loop backed during a sound effects is playing, the
background music would not to be played since its volume was reset to 0
and not set by commands correctly.

Fix:
Any commands are processed correctly even if the channel is muted.
In other words, when a channel is muted, commands of the channel are
processed correctly but are not applied / output to PSG registers
because it is muted.