fix!: removed or replaced __critical with DI/EI.
The `__critical` block in SDCC is useful for implementing critical
sections. It saves the enabled/disabled state of the interrupts before
executing the code in the block, and restores that state after the code
is executed.

However, on the Z80, this may not work as expected because the `LD A,I`
instruction is buggy. (i.e. Interrupts may be disabled after the
`__critical` block even if it had been enabled before.)

Therefore, we decided to use the pair `__asm__("di")` and
`__asm__("ei")` instead, or remove `__critical`.