fix(vgm2la): fixed VGM version compatibility issue
`vgm2la` does not support "Dual Chip Support" feature of VGM 1.70 or
later. So it is fixed to fail with error if the version was too newer.

Also, modified to skip VGM command 0x90..0x95 (DAC Stream Control
Write). This is because an unnecessary VGM command 0x94 was found in
some of the VGM files exported by the Furnace tracker.