fix: #81 Added APPLCODE which follows CODE segment
Added code segment `APPLCODE` following code segment `CODE` and modified
`mk/build.mk` to force user application code to be placed in the
`APPLCODE` segment by default. This will place the libmsx library and
standard C library code in the CODE segment and the user application
code in the APPLCODE segment. So as long as the total size of the
library code used by the application is approximately 16 KiB or less, it
will work as expected.