fix!: #80 fixed `mk/build.mk` and `tools/ridxc`
In `mk/build.mk`, the C source files were not passed to `ridxc`, so `ridxc` could not know if `BANK1`, `BANK2`, ... were used or not.

NOTE:
With this fix, `libmsx_resources.c` is now generated in the `obj` folder
instead of the `src` folder. So, if you have an old `src/libmsx_resources.c` in your application source tree, please remove
it manually first. Then run `make clean` or `make all` and it will work
fine again.